Create v/p distinction in algorithms
Algorithms have been prefixed with a 'v' or a 'p' to indicate that it is best suited for virtual or physical memory management. Physical just means that the memory that is to be operated on is mapped in. It can be read from and written to by the algorithm. Virtual means the opposite, the algorithm cannot access the memory it is operating on, so it must employ additional metadata structures.