Atomic pfreelist, VMM, vbuddy ialloc & ifree
The pfreelist has been implemented atomically, for single object allocations, such as to not cause a deadlock in the event that the page fault handler needs to map in some memory. There is a new per process VMM, and with this changes have been made to allow for the ialloc and ifree functions to be specified - as the internal allocator should  only be used if there is no other allocator available (i.e. when memory management is being initialized).