Lockless algos?, realloc, slab_expand
I want to make the pfreelist and vbuddy implementations lockless since that would be quite nice. I am just not entirely sure on how to do that. I have removed the realloc functions from the internal and general allocators as they are not needed. The functions to expand both of these allocators has also been implemented. Next I am planning on removing the ability to link pfreelists. This is not something that needs to be implemented specifically in a memory management algorithm, but rather something that can be made into its own abstraction: memory banks.