Pfreelist: check for empty lists
Allocations on hardware failed as an attempt was made to allocate from a freelist which had no more free objects. Moreover, the `head` member of the freelist struct was convert to an HHDM pointer in the pmm_convert_banks_to_hhdm function - causing it to no longer be NULL. This has also been rectified by adding a check.