vwatermark/pmm: create/bias checks
I have created the vwatermark algorithm. This is so that a watermark
allocator may be used later on in the VMM. It has the 'V' designation
as the implementation of the algorithm does not directly operate on
the memory it needs to manage. The PMM now does the calculation for
PMM_BIAS_COUNT and checks it against the other bias arrays to ensure
that they are all of equal length at compile time and does a runtime
check to use only the number of biases that fit into the address width.