PMM/pfreelist/pbuddy: use pbuddies/atomics/init
The PMM is now able to initialize and use pbuddies to preform
allocations of powers of two that have not been specified in the bias
list. I have added synchronization to the pfreelists and made the
allocation and freeing of an address atomic. The synchronization is key
to make the atomic allocations work as they delay processors. This
ensures that checks done before the atomic instruction are valid when
the atomic is preformed. The spinlock is so that the list of metas can
be re-ordered. I have begun work on the pbuddy implementation by
creating the initialization function and the function to allocate a new
meta. Next will be the allocation and freeing functions.