algo: removed vbuddy, implemented pbuddy
The old vbuddy algorithm has been replaced by the new pbuddy. This
new pbuddy algorithm requires the memory it is managing to be
present (hence the 'p' designation). This algorithm links together free
nodes making it much quicker to find a free node to allocate and split
and then to release nodes. Sizing information is kept in a separate
array whose elements represent objects sized as the lowest power of two
the algorithm is to represent (1 << PMM_BUDDY_LOWEST_EXPONENT). This
additional metadata allows for the tracking of the size of each
allocation / node letting the allocator not lose track of larger
portions of memory. A trend with these 'p' algorithms is to have three
structures: ARC_Algo, ARC_AlgoMeta, Arc_AlgoNode. The first keep track
of a linked list of metas, the second holds the data the algorithm
actually needs, and the third represents a physical node within the
range of memory being managed.