PMM: fixes to fast page pool, buddies, ordering
The fast page pool was experiencing some trouble as it was returning the
same page for repeated allocations (without freeing). This was due to
improper initialization. That has now been corrected so that the page
starting at the very last address is no longer included and the
freelists are properly linked together beginning to end. Moreover, there
was an issue with the fast page pool's freeing function. The order for
linking the newly freed page into the pool was reversed that is also
fixed. In the general allocation function there was an issue with the
improper PBuddy being used, this is because the index used was `i` and
not t_size_exp. Finally there was an issue with the ordering of
invokations to the various managers. This has been fixed. In the
allocation function it's: fast page pool, pfreelists, pbuddies and in
the free function it's: pbuddies, pfreelists, fast page pool.