pfreelist/pbuddy: fix meta insertion/mem getting
In the pfreelist and pbuddy implementations there was a race condition
between the insertion of a new meta structure and the re-ordering that
happens during allocation. This has been resolved by using the order
lock during the insertion of a new meta instead of doing it with an
atomic exchange. It may be valuable to try and get rid of this lock, or
to make it have a finer scope to reduce possible contention. The
pbuddy_alloc function, upon receiving NULL from the pbuddy_acquire
function, it attempts to create a new pbuddy meta structure for another
region of memory and rerun the allocation function.