vwatermark: freeing
For pwatermark freeing could not be implemented as addition
metadata (that cannot be gotten) is needed. In the vwatermark allocator,
the general purpose allocator will be available; therefore, additional
metadata can be stored to keep track of allocated regions and free
regions. A function to merge free regions has also been implemented to
hopefully mitigate fragmentation.